added context for prop management in Create Poll screen
